refactor(dashboard): use shared sortContracts util and SelectMarkets for add modal
